DICTIONARY SEARCH METHOD [MODIFICATION]
The method for searching words in the dictionary is now much more efficient. 
dictionary.txt is now organized in 27 lines, where each line contains all the words that start with a specific letter, separated by a ",".
The searchInDictionary method now searches for the line with the corresponding letter and then applies a binarySearch in the line.